bufhrt: new option --dsyncs-per-second
The --dsync option turned out to cause difficult timing conditions,
the time for syncing depends on drivers and hardware controlers and
can vary quite a bit on different calls.

This worked only with small values of --loops-per-second which in turn
causes large blocks to be writtten that don't fit into L1 cache  (or
very slow operation).

The new option --dsyncs-per-second gives more control over syncing,
independent of the --loops-per-second setting. A good starting point
for experiments is --loops-per-second=2000 and --dsyncs-per-second=100.