New program 'cleancp'
This new standalone program can be used instead of scripts/improvefile.

    cleancp origfile copiedfile

Call it without arguments to see the optional arguments, the defaults
should almost always be fine.

In contrast to the use of bufhrt in improvefile this program
  - uses mmap to access the input and output files
    (for reading it accesses directly the system page cache)
  - operates with the system page size
  - preallocates the blocks for the output file
  - syncronizes the output regularly

In tests the precise timing worked up to 3000-4000 pages per second
(the default is 2000 which should be always fine if the page size
is 4098 bytes).

Also a syncronization of the output every 32 blocks seems to work fine
(this is the default).