Small (private) utility by4.
"by4" is a filter for a stereo stream of 4-byte samples. It outputs
every fourth sample of each channel, reducing the sample rate by a
factor of four. Of course, the content must be low pass filtered before
(otherwise use resample_soxr).

Optionally, there can be an integer argument that has the same meaning
as --number-copies in other programs.