Add statistics loop to highrestest
added argument -h/--help

highrestest intv shift  (defaults intv = 500, shift = 100000)

Starts two loops for 10 seconds which sleeps until breakpoints
with 1ms distance and measures the times between the breakpoints
(of course, up to the systems resolution which will be less than
the claimed 1ns).

In first loop the breakpoints are just reached by calls
of clock_nanosleep().

In the second loop the time is measured after clock_nanosleep()
and then a simple loop is executed that should be ready at
the expected time + shift ns (so all breakpoints are shifted
by shift ns).